refactor: 테스트 파라미터 메소드를 별도 파일로 분리
