+++ src/test/java/domain/CoordinateTest.java
@@ -2,12 +2,8 @@ package domain;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.Arrays;
-import java.util.stream.Stream;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -18,7 +14,7 @@ public class CoordinateTest {
 
     @ParameterizedTest
     @DisplayName("정상적으로 자르는가")
-    @MethodSource("numbersProvider")
+    @MethodSource("parameter.TestParameter#numbersProvider")
     public void processTest(String string, Coordinate result) {
         assertTrue(new Coordinate(string).isEquals(result));
     }
@@ -28,13 +24,5 @@ public class CoordinateTest {
     @ValueSource(strings = {"(0,2)", "(1,25)"})
     public void processCheckExceptionTest(String string) {
         assertThrows(OutOfRangeException.class, () -> new Coordinate(string));
-    }
-
-    private static Stream<Arguments> numbersProvider() {
-        return Stream.of(
-            Arguments.of("(1,2)", new Coordinate(Arrays.asList(1, 2))),
-            Arguments.of("(24,1)", new Coordinate(Arrays.asList(24, 1))),
-            Arguments.of("(13,22)", new Coordinate(Arrays.asList(13, 22)))
-        ).map(arguments -> Arguments.of(arguments.get()[0], arguments.get()[1]));
     }
 }
